+++ CHANGELOG.md
@@ -36,6 +36,8 @@ We use the following categories for changes:
 
 ### Changed
 
+- deploy kube-state-metrics and custom metrics scraping by default [#sc-10432]
+
 ### Fixed
 
 ### Removed